+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog Project X
 
 #### Em desenvolvimento
 Novo
+- Integração de API ChatGPT
+- Nova interface
 - Menu de ajustes
 
 Correções/Melhorias